requirements.txt
tensorflow
tensorflow_hub
matplotlib
six
numpy
Pillow
absl-py
pathlib
IPython
numpy